+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Added parsing of `logical()` and `string()`
 * Reformatting of `matlab2r()` listed output
+* `matlab2r()` now accepts strings of MATLAB code as input
 
 # matlab2r 1.3.0